Update Script for cleaning mp1 data.R
Added in the labels for number of days bleeding (28/10) - ran the code on flo_26oct to make sure the code ran with no hitches and updated mp1.Rda in my OneDrive. This dataset is still lacking the YPA/YPB variables, smoking & alcohol variables need checking and other variables that have been put in the dataset that I didn't choose (action points for meeting 29/10). Updated the beginning of the script so when it is run start to finish, it will clear the data stored in the global environment, make a copy of the raw data leaving the original intact, change all the column names and variable category labels, then resave mp1.Rda with most up-to-date labels. This will be easy to adapt following variable changes that will arise from meeting 2.